+++ migrations/002_create_transactions_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE transactions (
     datetime TIMESTAMPTZ NOT NULL,
     user_id INTEGER NOT NULL,
-    amount FLOAT NOT NULL,
+    amount DECIMAL NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users (id)
 );
